Conflicts:
	WebContent/WEB-INF/view/service/board/main.jsp
	src/controller/BoardController.java

+++ src/controller/BoardController.java
@@ -520,31 +520,57 @@ public class BoardController {
 
 	}
 
-	@RequestMapping(path = "/boardDel", method = RequestMethod.GET)
-	public String Delete(@RequestParam String id, @RequestParam String menu, Model model, HttpSession session) {
-		System.out.println(id);
-		boolean rst = boardService.delete(id);
+	@RequestMapping("/change")
+	public String Change(Model model, @RequestParam(required = false) String no) {
+		// 이걸 no에 해당하는 데이터를 가지고 오고
+		model.addAttribute("title", "aaaa");
+		model.addAttribute("content", "dasdsadsadsad");
 
-		if (rst) {
-			model.addAttribute("find", boardService.findAll());
-			model.addAttribute("succ", "게시물이 삭제 되었습니다.");
+		return "/board/change";
+	}
+
+	
+	@RequestMapping(path = "/boardDel", method = RequestMethod.POST)
+	public String Delete(@RequestParam String id, @RequestParam String menu, Model model,HttpSession session) {
+		
+		Board board = boardService.find(id);
+		if(!board.getWriter().equals(session.getAttribute("logon"))) {
+			model.addAttribute("succ", "게시물 삭제는 작성자만 가능합니다.");
 			model.addAttribute("menu", menu);
 			model.addAttribute("page", 1);
 			return "redirect:/board/main";
-		} else {
-			model.addAttribute("read", boardService.find(id));
-			return "read_logon";
+		}else {
+			boolean rst = boardService.delete(id);
+			if (rst) {
+				model.addAttribute("find", boardService.findAll());
+				model.addAttribute("succ", "게시물이 삭제 되었습니다.");
+				model.addAttribute("menu", menu);
+				model.addAttribute("page", 1);
+				return "redirect:/board/main";
+			} else {
+				model.addAttribute("read", boardService.find(id));
+				return "read_logon";
+			}
 		}
 
 	}
-
-	@RequestMapping(path = "/Bedit", method = RequestMethod.GET)
-	public String GetEdit(@RequestParam String id, Model model) {
-
-		model.addAttribute("station", stationservice.readAllStation());
-		model.addAttribute("read", boardService.find(id));
-
-		return "board_edit";
+	
+	@RequestMapping(path="/bedit", method = RequestMethod.POST)
+	public String GetEdit(@RequestParam String id, @RequestParam String menu, Model model, HttpSession session) {
+		
+		Board board = boardService.find(id);
+		if(!board.getWriter().equals(session.getAttribute("logon"))) {
+			model.addAttribute("succ", "게시물 수정은 작성자만 가능합니다.");
+			model.addAttribute("menu", menu);
+			model.addAttribute("page", 1);
+			return "redirect:/board/main";
+		}else {
+			model.addAttribute("station", stationservice.readAllStation());
+			model.addAttribute("read", boardService.find(id));
+			
+			return "board_edit";
+		}
+		
 	}
 
 	@RequestMapping(path = "/Bedit", method = RequestMethod.POST)
